+++ src/main/resources/default/mc/sourceParameter_mc.xml
@@ -6,35 +6,35 @@
     sourceZdKey="MSimSourcePos.fZd"
     sourceAzKey="MSimSourcePos.fAz"
 />
-<fact.features.source.AntiSourcePosition
+<fact.features.source.OffPosition
     sourcePositionKey="sourcePosition"
-    numberOfAntiSourcePositions="5"
-    antiSourcePositionId="1"
-    outputKey="antiSourcePosition_1"
+    numberOfOffPositions="5"
+    offPositionId="1"
+    outputKey="offPosition1"
 />
-<fact.features.source.AntiSourcePosition
+<fact.features.source.OffPosition
     sourcePositionKey="sourcePosition"
-    numberOfAntiSourcePositions="5"
-    antiSourcePositionId="2"
-    outputKey="antiSourcePosition_2"
+    numberOfOffPositions="5"
+    offPositionId="2"
+    outputKey="offPosition2"
 />
-<fact.features.source.AntiSourcePosition
+<fact.features.source.OffPosition
     sourcePositionKey="sourcePosition"
-    numberOfAntiSourcePositions="5"
-    antiSourcePositionId="3"
-    outputKey="antiSourcePosition_3"
+    numberOfOffPositions="5"
+    offPositionId="3"
+    outputKey="offPosition3"
 />
-<fact.features.source.AntiSourcePosition
+<fact.features.source.OffPosition
     sourcePositionKey="sourcePosition"
-    numberOfAntiSourcePositions="5"
-    antiSourcePositionId="4"
-    outputKey="antiSourcePosition_4"
+    numberOfOffPositions="5"
+    offPositionId="4"
+    outputKey="offPosition4"
 />
-<fact.features.source.AntiSourcePosition
+<fact.features.source.OffPosition
     sourcePositionKey="sourcePosition"
-    numberOfAntiSourcePositions="5"
-    antiSourcePositionId="5"
-    outputKey="antiSourcePosition_5"
+    numberOfOffPositions="5"
+    offPositionId="5"
+    outputKey="offPosition5"
 />
 <fact.features.source.Alpha
     deltaKey="delta"
@@ -51,61 +51,61 @@
 <fact.features.source.Alpha
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_1"
+    sourcePositionKey="offPosition1"
     outputKey="Alpha_Off_1"
 />
 <fact.features.source.Distance
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_1"
+    sourcePositionKey="offPosition1"
     outputKey="Distance_Off_1"
 />
 <fact.features.source.Alpha
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_2"
+    sourcePositionKey="offPosition2"
     outputKey="Alpha_Off_2"
 />
 <fact.features.source.Distance
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_2"
+    sourcePositionKey="offPosition2"
     outputKey="Distance_Off_2"
 />
 <fact.features.source.Alpha
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_3"
+    sourcePositionKey="offPosition3"
     outputKey="Alpha_Off_3"
 />
 <fact.features.source.Distance
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_3"
+    sourcePositionKey="offPosition3"
     outputKey="Distance_Off_3"
 />
 <fact.features.source.Alpha
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_4"
+    sourcePositionKey="offPosition4"
     outputKey="Alpha_Off_4"
 />
 <fact.features.source.Distance
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_4"
+    sourcePositionKey="offPosition4"
     outputKey="Distance_Off_4"
 />
 <fact.features.source.Alpha
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_5"
+    sourcePositionKey="offPosition5"
     outputKey="Alpha_Off_5"
 />
 <fact.features.source.Distance
     deltaKey="delta"
     cogKey="cog"
-    sourcePositionKey="antiSourcePosition_5"
+    sourcePositionKey="offPosition5"
     outputKey="Distance_Off_5"
 />
 <fact.features.source.CosDeltaAlpha
@@ -115,31 +115,31 @@
     outputKey="CosDeltaAlpha"
 />
 <fact.features.source.CosDeltaAlpha
-    sourcePositionKey="antiSourcePosition_1"
+    sourcePositionKey="offPosition1"
     cogKey="cog"
     deltaKey="delta"
     outputKey="CosDeltaAlpha_Off_1"
 />
 <fact.features.source.CosDeltaAlpha
-    sourcePositionKey="antiSourcePosition_2"
+    sourcePositionKey="offPosition2"
     cogKey="cog"
     deltaKey="delta"
     outputKey="CosDeltaAlpha_Off_2"
 />
 <fact.features.source.CosDeltaAlpha
-    sourcePositionKey="antiSourcePosition_3"
+    sourcePositionKey="offPosition3"
     cogKey="cog"
     deltaKey="delta"
     outputKey="CosDeltaAlpha_Off_3"
 />
 <fact.features.source.CosDeltaAlpha
-    sourcePositionKey="antiSourcePosition_4"
+    sourcePositionKey="offPosition4"
     cogKey="cog"
     deltaKey="delta"
     outputKey="CosDeltaAlpha_Off_4"
 />
 <fact.features.source.CosDeltaAlpha
-    sourcePositionKey="antiSourcePosition_5"
+    sourcePositionKey="offPosition5"
     cogKey="cog"
     deltaKey="delta"
     outputKey="CosDeltaAlpha_Off_5"
@@ -204,27 +204,27 @@
 	outputKey="Theta"
 />
 <fact.features.source.Theta
-	sourcePositionKey="antiSourcePosition_1"
+	sourcePositionKey="offPosition1"
 	reconstructedPositionKey="reconstructedSourcePositionOff1"
 	outputKey="Theta_Off_1"
 />
 <fact.features.source.Theta
-	sourcePositionKey="antiSourcePosition_2"
+	sourcePositionKey="offPosition2"
 	reconstructedPositionKey="reconstructedSourcePositionOff2"
 	outputKey="Theta_Off_2"
 />
 <fact.features.source.Theta
-	sourcePositionKey="antiSourcePosition_3"
+	sourcePositionKey="offPosition3"
 	reconstructedPositionKey="reconstructedSourcePositionOff3"
 	outputKey="Theta_Off_3"
 />
 <fact.features.source.Theta
-	sourcePositionKey="antiSourcePosition_4"
+	sourcePositionKey="offPosition4"
 	reconstructedPositionKey="reconstructedSourcePositionOff4"
 	outputKey="Theta_Off_4"
 />
 <fact.features.source.Theta
-	sourcePositionKey="antiSourcePosition_5"
+	sourcePositionKey="offPosition5"
 	reconstructedPositionKey="reconstructedSourcePositionOff5"
 	outputKey="Theta_Off_5"
 />
