+++ src/main/resources/default/settings_mc.properties
@@ -55,7 +55,7 @@ pedestal_param=numPixelInPedestal,pedestalSize,pedestalTimespread,ped_mean_mean,
 
 fluctuations_param=fluct_mean_mean,fluct_mean_max,fluct_mean_min,fluct_mean_kurtosis,fluct_mean_variance,fluct_mean_skewness,fluct_mean_median,fluct_mean_p25,fluct_mean_p75,fluct_median_mean,fluct_median_max,fluct_median_min,fluct_median_kurtosis,fluct_median_variance,fluct_median_skewness,fluct_median_median,fluct_median_p25,fluct_median_p75,fluct_var_mean,fluct_var_max,fluct_var_min,fluct_var_kurtosis,fluct_var_variance,fluct_var_skewness,fluct_var_median,fluct_var_p25,fluct_var_p75,fluct_std_mean,fluct_std_max,fluct_std_min,fluct_std_kurtosis,fluct_std_variance,fluct_std_skewness,fluct_std_median,fluct_std_p25,fluct_std_p75,fluct_sum_mean,fluct_sum_max,fluct_sum_min,fluct_sum_kurtosis,fluct_sum_variance,fluct_sum_skewness,fluct_sum_median,fluct_sum_p25,fluct_sum_p75
 
-sourcePos=sourcePositionX,sourcePositionY,antiSourcePosition_*X,antiSourcePosition_*Y,reconstructedSourcePosition*X,reconstructedSourcePosition*Y
+sourcePos=sourcePositionX,sourcePositionY,offPosition*X,offPosition*Y,reconstructedSourcePosition*X,reconstructedSourcePosition*Y
 alpha=Alpha,Alpha_Off_1,Alpha_Off_2,Alpha_Off_3,Alpha_Off_4,Alpha_Off_5
 distance=Distance,Distance_Off_1,Distance_Off_2,Distance_Off_3,Distance_Off_4,Distance_Off_5
 cosdeltaalpha=CosDeltaAlpha,CosDeltaAlpha_Off_1,CosDeltaAlpha_Off_2,CosDeltaAlpha_Off_3,CosDeltaAlpha_Off_4,CosDeltaAlpha_Off_5
